refactor: move parsing to its own module (config-2) (#2875)
